+++ app/src/main/java/io/lunarlogic/aircasting/screens/dashboard/SessionsController.kt
@@ -62,7 +62,6 @@ abstract class SessionsController(
     open fun onPause() {
         unregisterSessionsObserver()
         mViewMvc.unregisterListener(this)
-        EventBus.getDefault().unregister(this)
     }
 
     protected fun startNewSession(sessionType: Session.Type) {
@@ -181,15 +180,5 @@ abstract class SessionsController(
         }
         val chooser = Intent.createChooser(sendIntent, context?.getString(R.string.share_link))
         context?.startActivity(chooser)
-    }
-
-    @Subscribe
-    fun onMessageEvent(event: NewMeasurementEvent) {
-        Thread.sleep(5000) //todo: this will be removed after code review and test
-        GlobalScope.launch {
-            withContext(Dispatchers.Main) {
-                mViewMvc.hideLoaderFor(event.deviceId!!)
-            }
-        }
     }
 }
